Create convience methods
Add documentation onto interfaces
