+++ SImpleRabbit.NetCore.Publisher/SimpleRabbit.NetCore.Publisher.csproj
@@ -8,7 +8,7 @@
     <PackageProjectUrl>https://github.com/blckmn/SimpleRabbit</PackageProjectUrl>
     <RepositoryUrl>https://github.com/blckmn/SimpleRabbit</RepositoryUrl>
 
-    <Description>An easy wrapper for the RabbitMQ client that allows inclusion in DotNetCore projects. Can be used with the standard dependency injection provided in DotNetCore. Provides a HostedService implementation for subscription based DotNetCore services.</Description>
+    <Description>An easy wrapper for the RabbitMQ client that allows inclusion in DotNetCore projects. Provides a factory that can be used to have multiple rabbit configurations</Description>
     <Copyright>Copyright © BLCKMN 2018</Copyright>
     <Summary>A simple DotNetCore (2.1) HostedService Subscriber wrapper for the RabbitMQ client.</Summary>
     <NeutralLanguage>en-US</NeutralLanguage>
@@ -20,6 +20,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\SimpleRabbit.NetCore\SimpleRabbit.NetCore.csproj" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.1.1" />
   </ItemGroup>
 
 </Project>
