--- SimpleRabbit.NetCore/Service/MessageHandler.cs
@@ -1,8 +1,0 @@
-﻿namespace SimpleRabbit.NetCore
-{
-    public interface IMessageHandler
-    {
-        bool CanProcess(string tag);
-        bool Process(BasicMessage message);
-    }
-}
